split global/challenge config parsing into separate files
